refactor: Apple Token Service 분리
