+++ src/main/java/Skeep/backend/auth/apple/service/AppleTokenUtil.java
@@ -1,0 +1,41 @@
+package Skeep.backend.auth.apple.service;
+
+import Skeep.backend.auth.exception.OAuthErrorCode;
+import Skeep.backend.global.exception.BaseException;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.jsonwebtoken.Claims;
+import io.jsonwebtoken.Jwts;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+
+import java.security.PublicKey;
+import java.util.Base64;
+import java.util.Map;
+
+@Component
+@RequiredArgsConstructor
+public class AppleTokenUtil {
+    private static final String TOKEN_VALUE_DELIMITER = "\\.";
+    private static final int HEADER_INDEX = 0;
+    private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
+
+    public Map<String, String> parseHeaders(String token) {
+        String encodedHeader = token.split(TOKEN_VALUE_DELIMITER)[HEADER_INDEX];
+        String decodedHeader = new String(Base64.getUrlDecoder().decode(encodedHeader));
+
+        try {
+            return OBJECT_MAPPER.readValue(decodedHeader, Map.class);
+        } catch (JsonProcessingException e) {
+            throw new BaseException(OAuthErrorCode.CANNOT_JSON_PROCESS);
+        }
+    }
+
+    public Claims getTokenClaims(String token, PublicKey publicKey) {
+        return Jwts.parserBuilder()
+                .setSigningKey(publicKey)
+                .build()
+                .parseClaimsJws(token)
+                .getBody();
+    }
+}
