+++ src/test/java/Skeep/backend/global/ControllerTest.java
@@ -4,6 +4,7 @@ import Skeep.backend.auth.apple.controller.AppleController;
 import Skeep.backend.auth.apple.service.AppleAuthClient;
 import Skeep.backend.auth.apple.service.ApplePublicKeyGenerator;
 import Skeep.backend.auth.apple.service.AppleService;
+import Skeep.backend.auth.apple.service.AppleTokenUtil;
 import Skeep.backend.global.security.config.SecurityConfig;
 import Skeep.backend.global.security.filter.JwtAuthenticationFilter;
 import Skeep.backend.global.security.filter.JwtExceptionFilter;
@@ -65,6 +66,9 @@ public abstract class ControllerTest {
 
     @MockBean
     protected JwtUtil jwtUtil;
+
+    @MockBean
+    protected AppleTokenUtil appleTokenUtil;
 
     @MockBean
     protected UserRepository userRepository;
